Heavy lifting with memory management.
- Stemming and FSA traversal require a constant memory pool and do not allocate anything dynamically.
- Custom codepage encoders and decoders used (to avoid temporary strings).

IStemmer API changed. A list of WordData is returned (possibly empty). Each WordData has a binary (byte[]) entry for a stem and tag and the possibility of converting these to a CharSequence.

The results are funny (seems to be slower than before). I am investigating what's the cause of this (besides low-level optimizations at JVM level).