Marcin, I'm switching to plain BSD license as stated here: http://www.opensource.org/licenses/bsd-license.php
This is a follow up of the discussion that took place on Apache legal mailing lists. The
"no-patents" clause that we added to BSD does not make much sense from the legal point of
view and is not entirely clear. As advised, it is better to go with plain BSD and go to 
court should anyone try to file a patent that covers prior art.